changing test structure
