+++ google-datacatalog-mysql-connector/tests/google/datacatalog_connectors/mysql_/lineage_synchronizer/scrape/tableLineage_test.py
@@ -49,4 +49,3 @@ class TestTableLineage():
         result = tableLineageExtractor().extract(query)
         print(result)
         assert result == expected
-
