+++ google-datacatalog-mysql-connector/tests/google/datacatalog_connectors/mysql_/lineage_synchronizer/scrape/test_data/simpleUpdateTable/query.sql
@@ -1,7 +1,9 @@
 UPDATE
   cars C,
-  users U
+  users U,
+  (SELECT * FROM T1 where x > 1) as t
 SET
-  vendor = U.first_name
+  vendor = U.first_name,
+  x = y
 WHERE
   C.id = U.id;
